Rename Rotation to Orientation for better naming accuracy
