+++ src/main/java/org/spongepowered/api/CatalogTypes.java
@@ -34,7 +34,38 @@ import org.spongepowered.api.block.entity.BlockEntityType;
 import org.spongepowered.api.command.selector.SelectorType;
 import org.spongepowered.api.data.persistence.DataFormat;
 import org.spongepowered.api.data.persistence.DataTranslator;
-import org.spongepowered.api.data.type.*;
+import org.spongepowered.api.data.type.ArtType;
+import org.spongepowered.api.data.type.BannerPatternShape;
+import org.spongepowered.api.data.type.BoatType;
+import org.spongepowered.api.data.type.BodyPart;
+import org.spongepowered.api.data.type.CatType;
+import org.spongepowered.api.data.type.ComparatorMode;
+import org.spongepowered.api.data.type.DoorHinge;
+import org.spongepowered.api.data.type.DyeColor;
+import org.spongepowered.api.data.type.FoxType;
+import org.spongepowered.api.data.type.HandPreference;
+import org.spongepowered.api.data.type.HandType;
+import org.spongepowered.api.data.type.HorseColor;
+import org.spongepowered.api.data.type.HorseStyle;
+import org.spongepowered.api.data.type.InstrumentType;
+import org.spongepowered.api.data.type.LlamaType;
+import org.spongepowered.api.data.type.MooshroomType;
+import org.spongepowered.api.data.type.NotePitch;
+import org.spongepowered.api.data.type.PandaGene;
+import org.spongepowered.api.data.type.ParrotType;
+import org.spongepowered.api.data.type.PickupRule;
+import org.spongepowered.api.data.type.PistonType;
+import org.spongepowered.api.data.type.PortionType;
+import org.spongepowered.api.data.type.ProfessionType;
+import org.spongepowered.api.data.type.RabbitType;
+import org.spongepowered.api.data.type.RailDirection;
+import org.spongepowered.api.data.type.SkinPart;
+import org.spongepowered.api.data.type.SlabPortion;
+import org.spongepowered.api.data.type.StairShape;
+import org.spongepowered.api.data.type.ToolType;
+import org.spongepowered.api.data.type.TropicalFishShape;
+import org.spongepowered.api.data.type.WireAttachmentType;
+import org.spongepowered.api.data.type.WoodType;
 import org.spongepowered.api.effect.particle.ParticleOption;
 import org.spongepowered.api.effect.particle.ParticleType;
 import org.spongepowered.api.effect.potion.PotionEffectType;
@@ -45,22 +76,24 @@ import org.spongepowered.api.entity.ai.goal.GoalExecutorType;
 import org.spongepowered.api.entity.ai.goal.GoalType;
 import org.spongepowered.api.entity.attribute.AttributeOperation;
 import org.spongepowered.api.entity.attribute.type.AttributeType;
+import org.spongepowered.api.entity.living.player.chat.ChatVisibility;
 import org.spongepowered.api.entity.living.player.gamemode.GameMode;
 import org.spongepowered.api.event.EventContextKey;
+import org.spongepowered.api.event.cause.entity.DismountType;
+import org.spongepowered.api.event.cause.entity.MovementType;
+import org.spongepowered.api.event.cause.entity.SpawnType;
 import org.spongepowered.api.event.cause.entity.damage.DamageModifierType;
 import org.spongepowered.api.event.cause.entity.damage.DamageType;
-import org.spongepowered.api.event.cause.entity.DismountType;
-import org.spongepowered.api.event.cause.entity.SpawnType;
-import org.spongepowered.api.event.cause.entity.MovementType;
 import org.spongepowered.api.fluid.FluidType;
 import org.spongepowered.api.item.FireworkShape;
 import org.spongepowered.api.item.ItemType;
 import org.spongepowered.api.item.enchantment.EnchantmentType;
 import org.spongepowered.api.item.inventory.ContainerType;
-import org.spongepowered.api.item.inventory.equipment.EquipmentType;
 import org.spongepowered.api.item.inventory.equipment.EquipmentGroup;
+import org.spongepowered.api.item.inventory.equipment.EquipmentType;
 import org.spongepowered.api.item.inventory.query.QueryType;
 import org.spongepowered.api.item.recipe.crafting.CraftingRecipe;
+import org.spongepowered.api.placeholder.PlaceholderParser;
 import org.spongepowered.api.registry.CatalogRegistry;
 import org.spongepowered.api.registry.GameRegistry;
 import org.spongepowered.api.scoreboard.CollisionRule;
@@ -68,22 +101,20 @@ import org.spongepowered.api.scoreboard.Visibility;
 import org.spongepowered.api.scoreboard.criteria.Criterion;
 import org.spongepowered.api.scoreboard.displayslot.DisplaySlot;
 import org.spongepowered.api.scoreboard.objective.displaymode.ObjectiveDisplayMode;
+import org.spongepowered.api.service.ban.BanType;
 import org.spongepowered.api.service.economy.Currency;
 import org.spongepowered.api.service.economy.transaction.TransactionType;
-import org.spongepowered.api.placeholder.PlaceholderParser;
 import org.spongepowered.api.statistic.Statistic;
 import org.spongepowered.api.statistic.StatisticCategory;
-import org.spongepowered.api.entity.living.player.chat.ChatVisibility;
-import org.spongepowered.api.service.ban.BanType;
-import org.spongepowered.api.util.rotation.Rotation;
-import org.spongepowered.api.world.dimension.DimensionType;
-import org.spongepowered.api.world.gen.GeneratorModifierType;
-import org.spongepowered.api.world.portal.PortalType;
+import org.spongepowered.api.util.orientation.Orientation;
 import org.spongepowered.api.world.SerializationBehavior;
 import org.spongepowered.api.world.WorldArchetype;
 import org.spongepowered.api.world.biome.BiomeType;
 import org.spongepowered.api.world.biome.VirtualBiomeType;
 import org.spongepowered.api.world.difficulty.Difficulty;
+import org.spongepowered.api.world.dimension.DimensionType;
+import org.spongepowered.api.world.gen.GeneratorModifierType;
+import org.spongepowered.api.world.portal.PortalType;
 import org.spongepowered.api.world.teleport.TeleportHelperFilter;
 import org.spongepowered.api.world.weather.Weather;
 
@@ -237,7 +268,7 @@ public final class CatalogTypes {
 
     public static final Class<RailDirection> RAIL_DIRECTION = RailDirection.class;
 
-    public static final Class<Rotation> ROTATION = Rotation.class;
+    public static final Class<Orientation> ROTATION = Orientation.class;
 
     public static final Class<SelectorType> SELECTOR_TYPE = SelectorType.class;
