+++ src/main/java/org/spongepowered/api/data/Keys.java
@@ -119,7 +119,6 @@ import org.spongepowered.api.entity.living.Ageable;
 import org.spongepowered.api.entity.living.Agent;
 import org.spongepowered.api.entity.living.ArmorStand;
 import org.spongepowered.api.entity.living.Bat;
-import org.spongepowered.api.entity.living.Human;
 import org.spongepowered.api.entity.living.Humanoid;
 import org.spongepowered.api.entity.living.Living;
 import org.spongepowered.api.entity.living.animal.Animal;
@@ -216,7 +215,7 @@ import org.spongepowered.api.util.Color;
 import org.spongepowered.api.util.Direction;
 import org.spongepowered.api.util.RespawnLocation;
 import org.spongepowered.api.util.Ticks;
-import org.spongepowered.api.util.rotation.Rotation;
+import org.spongepowered.api.util.orientation.Orientation;
 import org.spongepowered.api.util.weighted.WeightedSerializableObject;
 import org.spongepowered.api.world.ServerLocation;
 import org.spongepowered.api.world.explosion.Explosion;
@@ -1859,6 +1858,11 @@ public final class Keys {
     public static final Supplier<Key<Value<Boolean>>> ON_GROUND = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Key.class, "on_ground");
 
     /**
+     * The {@link Orientation} of an {@link ItemFrame}.
+     */
+    public static final Supplier<Key<Value<Orientation>>> ORIENTATION = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Key.class, "orientation");
+
+    /**
      * The content of a {@link ItemTypes#WRITTEN_BOOK} {@link ItemStack}.
      *
      * <p>Use {@link Keys#PLAIN_PAGES} if you wish to inspect the contents
@@ -2105,11 +2109,6 @@ public final class Keys {
      * The time a {@link Ravager} is roaring.
      */
     public static final Supplier<Key<Value<Ticks>>> ROARING_TIME = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Key.class, "roaring_time");
-
-    /**
-     * The {@link Rotation} of an {@link ItemFrame}.
-     */
-    public static final Supplier<Key<Value<Rotation>>> ROTATION = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Key.class, "rotation");
 
     /**
      * The current saturation of a {@link Player}.
