+++ src/main/java/org/spongepowered/api/util/orientation/Orientations.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.util.rotation;
+package org.spongepowered.api.util.orientation;
 
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.entity.hanging.ItemFrame;
@@ -35,30 +35,30 @@ import java.util.function.Supplier;
  * such as an {@link ItemStack} within
  * an {@link ItemFrame}.
  */
-public final class Rotations {
+public final class Orientations {
 
     // SORTFIELDS:ON
 
-    public static final Supplier<Rotation> BOTTOM = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Rotation.class, "bottom");
+    public static final Supplier<Orientation> BOTTOM = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Orientation.class, "bottom");
 
-    public static final Supplier<Rotation> BOTTOM_LEFT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Rotation.class, "bottom_left");
+    public static final Supplier<Orientation> BOTTOM_LEFT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Orientation.class, "bottom_left");
 
-    public static final Supplier<Rotation> BOTTOM_RIGHT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Rotation.class, "bottom_right");
+    public static final Supplier<Orientation> BOTTOM_RIGHT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Orientation.class, "bottom_right");
 
-    public static final Supplier<Rotation> LEFT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Rotation.class, "left");
+    public static final Supplier<Orientation> LEFT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Orientation.class, "left");
 
-    public static final Supplier<Rotation> RIGHT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Rotation.class, "right");
+    public static final Supplier<Orientation> RIGHT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Orientation.class, "right");
 
-    public static final Supplier<Rotation> TOP = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Rotation.class, "top");
+    public static final Supplier<Orientation> TOP = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Orientation.class, "top");
 
-    public static final Supplier<Rotation> TOP_LEFT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Rotation.class, "top_left");
+    public static final Supplier<Orientation> TOP_LEFT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Orientation.class, "top_left");
 
-    public static final Supplier<Rotation> TOP_RIGHT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Rotation.class, "top_right");
+    public static final Supplier<Orientation> TOP_RIGHT = Sponge.getRegistry().getCatalogRegistry().provideSupplier(Orientation.class, "top_right");
 
     // SORTFIELDS:OFF
 
     // Suppress default constructor to ensure non-instantiability.
-    private Rotations() {
+    private Orientations() {
         throw new AssertionError("You should not be attempting to instantiate this class.");
     }
